refactor: 체크리스트 질문 받아오는 로직 수정
