+++ frontend/src/pages/NewChecklistPage/NewChecklistPage.tsx
@@ -29,34 +29,35 @@ const DefaultRoomInfo: RoomInfo = {
   realEstate: '방끗공인중개사',
 };
 
-const categoryTabs: Tab[] = [
+export const categoryTabs: Tab[] = [
+  { id: 0, name: '기본 정보' },
   {
     id: 1,
-    categoryName: '청결',
+    name: '청결',
   },
   {
     id: 2,
-    categoryName: '방 컨디션',
+    name: '방 컨디션',
   },
   {
     id: 3,
-    categoryName: '편의시설',
+    name: '편의시설',
   },
   {
     id: 4,
-    categoryName: '옵션',
+    name: '옵션',
   },
   {
     id: 5,
-    categoryName: '주거환경',
+    name: '주거환경',
   },
   {
     id: 6,
-    categoryName: '보안',
+    name: '보안',
   },
   {
     id: 7,
-    categoryName: '경제적',
+    name: '경제적',
   },
 ];
